Move `BitIndex` and `BitReduction` under `BitPack`
And export functions from `Clash.Class.BitPack` to make them more
discoverable.
