+++ clash-ghc/src-ghc/Clash/GHC/GHC2Core.hs
@@ -369,7 +369,7 @@ coreToTerm primMap unlocs = term
         go "GHC.Stack.withFrozenCallStack"     args
           | length args == 3
           = term (App (args!!2) (args!!1))
-        go "Clash.Class.BitPack.packXWith" args
+        go "Clash.Class.BitPack.Internal.packXWith" args
           | [_nTy,_aTy,_kn,f] <- args
           = term f
         go "Clash.Sized.BitVector.Internal.checkUnpackUndef" args
@@ -538,7 +538,7 @@ coreToTerm primMap unlocs = term
               | f == "GHC.Magic.noinline"               -> return (idTerm xType)
               | f == "GHC.Magic.lazy"                   -> return (idTerm xType)
               | f == "GHC.Magic.runRW#"                 -> return (runRWTerm xType)
-              | f == "Clash.Class.BitPack.packXWith"    -> return (packXWithTerm xType)
+              | f == "Clash.Class.BitPack.Internal.packXWith"    -> return (packXWithTerm xType)
               | f == "Clash.Sized.Internal.BitVector.checkUnpackUndef" -> return (checkUnpackUndefTerm xType)
               | f == "Clash.Magic.prefixName"
               -> return (nameModTerm C.PrefixName xType)
