+++ clash-prelude/src/Clash/Sized/Fixed.hs
@@ -142,8 +142,8 @@ import Clash.Class.Num            (ExtendingNum (..), SaturatingNum (..),
                                    boundedMul)
 import Clash.Class.Resize         (Resize (..))
 import Clash.Promoted.Nat         (SNat, natToNum, natToInteger)
-import Clash.Prelude.BitIndex     (lsb, msb, split)
-import Clash.Prelude.BitReduction (reduceAnd, reduceOr)
+import Clash.Class.BitPack.BitIndex (lsb, msb, split)
+import Clash.Class.BitPack.BitReduction (reduceAnd, reduceOr)
 import Clash.Sized.BitVector      (BitVector, (++#))
 import Clash.Sized.Signed         (Signed)
 import Clash.Sized.Unsigned       (Unsigned)
